+++ src/protocol/parts/resultset_metadata.rs
@@ -116,7 +116,7 @@ impl ResultSetMetadata {
         Ok(self.get(i)?.precision())
     }
 
-    pub(crate) fn parse<T: io::BufRead>(count: i32, rdr: &mut T) -> HdbResult<ResultSetMetadata> {
+    pub(crate) fn parse<T: io::BufRead>(count: usize, rdr: &mut T) -> HdbResult<ResultSetMetadata> {
         let mut rsm = ResultSetMetadata {
             fields: Vec::<FieldMetadata>::new(),
             names: VecMap::<String>::new(),
